move modern Icon
