+++ src/components/FloatingButtons/FloatingButtons.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 import SafeAreaView from 'react-native-safe-area-view';
 
 // Components
-import Icon from 'components/modern/Icon';
+import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
 
 // Utils
@@ -32,7 +32,7 @@ import { spacing } from 'utils/variables';
 
 // Types
 import type { ImageSource } from 'utils/types/react-native';
-import type { IconName } from 'components/modern/Icon';
+import type { IconName } from 'components/core/Icon';
 
 export type Item = {|
   title: string,
