+++ src/components/HistoryList/items/HistoryListItem.js
@@ -23,7 +23,7 @@ import styled from 'styled-components/native';
 import { SvgCssUri } from 'react-native-svg';
 
 // Components
-import Icon, { type IconName } from 'components/modern/Icon';
+import Icon, { type IconName } from 'components/core/Icon';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
 import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
