+++ src/components/inputs/CollectibleInput.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 
 // Components
 import CollectibleImage from 'components/CollectibleImage';
-import Icon from 'components/modern/Icon';
+import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
 
 // Utils
