+++ src/components/modern/Button.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 
 // Components
 import Text from 'components/core/Text';
-import Icon, { type IconName } from 'components/modern/Icon';
+import Icon, { type IconName } from 'components/core/Icon';
 
 // Utils
 import { fontStyles, appFont, spacing } from 'utils/variables';
