+++ src/components/modern/TextWithCopy.js
@@ -23,7 +23,7 @@ import Clipboard from '@react-native-community/clipboard';
 import styled from 'styled-components/native';
 
 // components
-import Icon from 'components/modern/Icon';
+import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
 import Toast from 'components/Toast';
