+++ src/screens/Menu/Settings/components/SettingsToggle.js
@@ -22,7 +22,7 @@ import * as React from 'react';
 import styled from 'styled-components/native';
 
 // Components
-import Icon, { type IconName } from 'components/modern/Icon';
+import Icon, { type IconName } from 'components/core/Icon';
 import Switcher from 'components/Switcher';
 import Text from 'components/core/Text';
