+++ src/screens/WalletConnect/CallRequest/TransactionRequestContent.js
@@ -32,7 +32,7 @@ import Image from 'components/Image';
 import LargeFiatTokenValueView from 'components/modern/LargeFiatTokenValueView';
 import Text from 'components/core/Text';
 import TransactionDeploymentWarning from 'components/other/TransactionDeploymentWarning';
-import Icon from 'components/modern/Icon';
+import Icon from 'components/core/Icon';
 
 // Constants
 import { CHAIN } from 'constants/chainConstants';
