+++ src/screens/WalletConnect/Home/components/DeployBanner.js
@@ -24,7 +24,7 @@ import LinearGradient from 'react-native-linear-gradient';
 import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
-import Icon from 'components/modern/Icon';
+import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
 
 // Hooks
